# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LastPiece API
  version: "1.0.0"
  description: |
    Серверная API. Аутентификация: Bearer JWT (accessToken),
    refreshToken хранится в HttpOnly cookie. Базовый URL зависит от PORT
    (по умолчанию 3000).
servers:
  - url: http://localhost:3000
    description: Default PORT=3000 (см. src/shared/config/index.ts)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: BadRequest
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: NotFound
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string, nullable: true }
        details:
          type: object
          nullable: true
          additionalProperties: true
...
    PublicSeller:
      type: object
      description: Только публичные поля продавца.
      properties:
        id: { type: string }
        companyName: { type: string }
        website: { type: string, format: uri, nullable: true }
    ProductPublicItem:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        photos:
          type: array
          items: { type: string, format: uri }
        price: { type: number }
        oldPrice: { type: number }
        discountPercent: { type: integer }
        isHot: { type: boolean }
        offerTypeId: { type: string }
        mainCategoryId: { type: string }
        subcategoryId: { type: string, nullable: true }
        rankScore: { type: number }
        views7d: { type: integer }
        favoritesCount: { type: integer }
        createdAt: { type: string, format: date-time }
        seller: { $ref: '#/components/schemas/PublicSeller' }
    ProductPublicListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ProductPublicItem' }
        nextCursor: { type: string, nullable: true }
    ProductDetail:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        photos:
          type: array
          items: { type: string, format: uri }
        price: { type: number }
        oldPrice: { type: number }
        discountPercent: { type: integer }
        isHot: { type: boolean }
        offerTypeId: { type: string }
        mainCategoryId: { type: string }
        subcategoryId: { type: string, nullable: true }
        externalUrl: { type: string, nullable: true }
        views7d: { type: integer }
        views30d: { type: integer }
        favoritesCount: { type: integer }
        clicksToExternal7d: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        seller: { $ref: '#/components/schemas/PublicSeller' }
    ProductSimilarItem:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        photos:
          type: array
          items: { type: string, format: uri }
        price: { type: number }
        oldPrice: { type: number }
        discountPercent: { type: integer }
        isHot: { type: boolean }
    ProductSimilarResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ProductSimilarItem' }
    ProductClickResponse:
      type: object
      properties:
        externalUrl: { type: string, nullable: true }
    SellerProductItem:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        price: { type: number }
        oldPrice: { type: number }
        status: { type: string, enum: [draft, active, archived] }
        createdAt: { type: string, format: date-time }
        discountPercent: { type: integer }
    SellerProductsListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/SellerProductItem' }
        nextCursor: { type: string, nullable: true }
    SellerProductCreateRequest:
      type: object
      required: [title, oldPrice, price, mainCategoryId, offerTypeId]
      properties:
        title: { type: string, minLength: 2, maxLength: 160 }
        description: { type: string, maxLength: 2000 }
        photos:
          type: array
          items: { type: string, format: uri }
          description: Максимум файлов определяется config.uploads.maxFiles
        oldPrice: { type: number, exclusiveMinimum: 0 }
        price: { type: number, minimum: 0 }
        mainCategoryId: { type: string, minLength: 8 }
        subcategoryId: { type: string, minLength: 8 }
        offerTypeId: { type: string, minLength: 8 }
        externalUrl: { type: string, format: uri }
        status: { type: string, enum: [draft, active] }
      description: |
        price должен быть меньше oldPrice. Поле status игнорируется сервером,
        продукт создается со статусом draft.
    SellerProductUpdateRequest:
      type: object
      description: |
        Частичное обновление. Если изменяются цены, сервер требует oldPrice > price
        и может пересчитать rankScore. status допускает только draft/active.
      properties:
        title: { type: string, minLength: 2, maxLength: 160 }
        description: { type: string, maxLength: 2000 }
        photos:
          type: array
          items: { type: string, format: uri }
        oldPrice: { type: number, exclusiveMinimum: 0 }
        price: { type: number, minimum: 0 }
        mainCategoryId: { type: string, minLength: 8 }
        subcategoryId: { type: string, minLength: 8 }
        offerTypeId: { type: string, minLength: 8 }
        externalUrl: { type: string, format: uri }
        status: { type: string, enum: [draft, active] }
    SellerProductDetail:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        photos:
          type: array
          items: { type: string, format: uri }
        oldPrice: { type: number }
        price: { type: number }
        discountPercent: { type: integer }
        mainCategoryId: { type: string }
        subcategoryId: { type: string, nullable: true }
        offerTypeId: { type: string }
        externalUrl: { type: string, nullable: true }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AdminProductListItem:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        status: { type: string, enum: [draft, pending, active, rejected, archived] }
        price: { type: number }
        oldPrice: { type: number }
        discountPercent: { type: integer }
        sellerId: { type: string }
        mainCategoryId: { type: string }
        subcategoryId: { type: string, nullable: true }
        offerTypeId: { type: string }
        isFeatured: { type: boolean }
        featuredUntil: { type: string, format: date-time, nullable: true }
        adminTags:
          type: array
          items: { type: string, enum: [best, popular, new] }
        rankScore: { type: number }
        promotionUntil: { type: string, format: date-time, nullable: true }
        rankAdminBoost: { type: number }
        createdAt: { type: string, format: date-time }
    AdminProductListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminProductListItem' }
        nextCursor: { type: string, nullable: true }
    AdminProductDetail:
      type: object
      description: Полный документ продукта + ссылки seller/category/subcategory/offerType.
      properties:
        id: { type: string }
        sellerId: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        photos:
          type: array
          items: { type: string, format: uri }
        oldPrice: { type: number }
        price: { type: number }
        mainCategoryId: { type: string }
        subcategoryId: { type: string, nullable: true }
        offerTypeId: { type: string }
        externalUrl: { type: string, nullable: true }
        status: { type: string, enum: [draft, pending, active, rejected, archived] }
        deletedAt: { type: string, format: date-time, nullable: true }
        moderationNote: { type: string, nullable: true }
        approvedAt: { type: string, format: date-time, nullable: true }
        rejectedAt: { type: string, format: date-time, nullable: true }
        moderatedBy: { type: string, nullable: true }
        views7d: { type: integer }
        views30d: { type: integer }
        favoritesCount: { type: integer }
        clicksToExternal7d: { type: integer }
        rankScore: { type: number }
        shuffleKey: { type: number }
        isFeatured: { type: boolean }
        featuredUntil: { type: string, format: date-time, nullable: true }
        adminTags:
          type: array
          items: { type: string, enum: [best, popular, new] }
        rankAdminBoost: { type: number }
        promotionUntil: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        seller:
          type: object
          nullable: true
          properties:
            id: { type: string }
            companyName: { type: string }
            status: { type: string }
        category:
          type: object
          nullable: true
          properties:
            id: { type: string }
            name: { type: string }
        subcategory:
          type: object
          nullable: true
          properties:
            id: { type: string }
            name: { type: string }
        offerType:
          type: object
          nullable: true
          properties:
            id: { type: string }
            name: { type: string }
            slug: { type: string }
...
paths:
...
  /api/v1/products:
    get:
      tags: [Products (Public)]
      summary: Публичная витрина
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: sort
          in: query
          schema: { type: string, enum: [rank, new, popular], default: rank }
        - name: category
          in: query
          schema: { type: string }
        - name: subcategory
          in: query
          schema: { type: string }
        - name: offerTypeSlug
          in: query
          schema: { type: string }
        - name: hot
          in: query
          schema: { type: integer }
        - name: cursor
          in: query
          schema:
            type: string
            description: base64 cursor зависит от sort (см. код)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductPublicListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/products/{id}:
    get:
      tags: [Products (Public)]
      summary: Карточка товара
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/products/similar/{id}:
    get:
      tags: [Products (Public)]
      summary: Похожие товары
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductSimilarResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/products/{id}/click:
    post:
      tags: [Products (Public)]
      summary: Зарегистрировать клик и вернуть externalUrl
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductClickResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------- SELLER / PRODUCTS --------------
  /api/v1/seller/products:
    get:
      tags: [Seller/Products]
      security: [{ bearerAuth: [] }]
      summary: Мой список товаров
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [all, draft, active, archived], default: all }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: cursor
          in: query
          schema:
            type: string
            description: base64 JSON {"id":"<ObjectId>"}
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SellerProductsListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Seller/Products]
      security: [{ bearerAuth: [] }]
      summary: Создать товар
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SellerProductCreateRequest' }
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/seller/products/{id}:
    get:
      tags: [Seller/Products]
      security: [{ bearerAuth: [] }]
      summary: Детали товара продавца
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SellerProductDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Seller/Products]
      security: [{ bearerAuth: [] }]
      summary: Обновить товар продавца
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SellerProductUpdateRequest' }
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Seller/Products]
      security: [{ bearerAuth: [] }]
      summary: Удалить товар (soft-delete)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Удалено (soft) }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/admin/products:
    get:
      tags: [Admin/Products]
      security: [{ bearerAuth: [] }]
      summary: Листинг товаров (расширенный фильтр)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [all, draft, pending, active, rejected, archived], default: all }
        - name: q
          in: query
          schema: { type: string }
        - name: sellerId
          in: query
          schema: { type: string }
        - name: categoryId
          in: query
          schema: { type: string }
        - name: subcategoryId
          in: query
          schema: { type: string }
        - name: offerTypeId
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: integer, enum: [1] }
        - name: hot
          in: query
          schema: { type: integer, enum: [1] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: cursor
          in: query
          schema:
            type: string
            description: base64 JSON {"id":"<ObjectId>"}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProductListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/products/{id}:
    get:
      tags: [Admin/Products]
      security: [{ bearerAuth: [] }]
      summary: Детали товара (полный документ + ссылки)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Детали
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProductDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Admin/Products]
      security: [{ bearerAuth: [] }]
      summary: Удалить товар (hard delete)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Удалено }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...